requirements.txt
tensorflow==2.13.0
streamlit==1.52.2
pillow
numpy
